merges typing and items in extract
