+++ dlt/common/typing.py
@@ -93,6 +93,7 @@ TVariantBase = TypeVar("TVariantBase", covariant=True)
 TVariantRV = Tuple[str, Any]
 VARIANT_FIELD_FORMAT = "v_%s"
 TFileOrPath = Union[str, os.PathLike, IO[Any]]
+TSortOrder = Literal["asc", "desc"]
 
 
 @runtime_checkable
