+++ dlt/extract/decorators.py
@@ -62,7 +62,7 @@ from dlt.extract.exceptions import (
 )
 from dlt.extract.incremental import IncrementalResourceWrapper
 
-from dlt.extract.typing import TTableHintTemplate
+from dlt.extract.items import TTableHintTemplate
 from dlt.extract.source import DltSource
 from dlt.extract.resource import DltResource, TUnboundDltResource
